requirements.txt
celery==5.5.3
django==5.2.4
django-environ==0.12.0
psycopg==3.2.9
redis==6.3.0
